+++ fsl/utils/path.py
@@ -23,6 +23,8 @@ paths.
    removeDuplicates
    uniquePrefix
    commonBase
+   wslpath
+   winpath
 """
 
 
@@ -30,6 +32,9 @@ import os.path as op
 import            os
 import            glob
 import            operator
+import            re
+
+from fsl.utils.platform import platform
 
 
 class PathError(Exception):
@@ -524,3 +529,58 @@ def commonBase(paths):
             return base
 
     raise PathError('No common base')
+
+
+def wslpath(winpath):
+    """
+    Convert Windows path (or a command line argument containing a Windows path)
+    to the equivalent WSL path (e.g. ``c:\\Users`` -> ``/mnt/c/Users``). Also supports
+    paths in the form ``\\wsl$\\(distro)\\users\\...``
+
+    :param winpath: Command line argument which may (or may not) contain a Windows path. It is assumed to be
+                    either of the form <windows path> or --<arg>=<windows path>. Note that we don't need to
+                    handle --arg <windows path> or -a <windows path> since in these cases the argument
+                    and the path will be parsed as separate entities.
+    :return: If ``winpath`` matches a Windows path, the converted argument (including the --<arg>= portion).
+                Otherwise returns ``winpath`` unchanged.
+    """
+    match = re.match(r"^(--[\w-]+=)?\\\\wsl\$[\\\/][^\\^\/]+(.*)$", winpath)
+    if match:
+        arg, path = match.group(1, 2)
+        if arg is None:
+            arg = ""
+        return arg + path.replace("\\", "/")
+
+    match = re.match(r"^(--[\w-]+=)?([a-zA-z]):(.+)$", winpath)
+    if match:
+        arg, drive, path = match.group(1, 2, 3)
+        if arg is None:
+            arg = ""
+        return arg + "/mnt/" + drive.lower() + path.replace("\\", "/")
+
+    return winpath
+
+
+def winpath(wslpath):
+    """
+    Convert a WSL-local filepath (for example ``/usr/local/fsl/``) into a path that can be used from
+    Windows.
+
+    If ``self.fslwsl`` is ``False``, simply returns ``wslpath`` unmodified
+    Otherwise, uses ``FSLDIR`` to deduce the WSL distro in use for FSL.
+    This requires WSL2 which supports the ``\\wsl$\`` network path.
+    wslpath is assumed to be an absolute path.
+    """
+    if not platform.fslwsl:
+        return wslpath
+    else:
+        match = re.match(r"^\\\\wsl\$\\([^\\]+).*$", platform.fsldir)
+        if match:
+            distro = match.group(1)
+        else:
+            distro = None
+
+        if not distro:
+            raise RuntimeError("Could not identify WSL installation from FSLDIR (%s)" % platform.fsldir)
+
+        return "\\\\wsl$\\" + distro + wslpath.replace("/", "\\")